refactor push router lifecycle modules
