+++ lib/llm/src/kv_router/push_router.rs
@@ -1,22 +1,15 @@
 // SPDX-FileCopyrightText: Copyright (c) 2024-2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-use std::{collections::HashSet, future::Future, sync::Arc};
+use std::sync::Arc;
 
 use anyhow::Result;
-use dynamo_kv_router::{
-    RouterConfigOverride,
-    indexer::RoutingDecisionHashes,
-    protocols::{BlockExtraInfo, RoutingConstraints, TokensWithHashes, WorkerId, WorkerWithDpRank},
-    scheduling::{RoutingEligibility, WorkerEligibilityError},
-};
+use dynamo_kv_router::protocols::{TokensWithHashes, WorkerWithDpRank};
 use dynamo_runtime::{
-    dynamo_nvtx_range,
-    error::{DynamoError, ErrorType as DynamoErrorType},
     metrics::frontend_perf::{STAGE_ROUTE, StageGuard},
     pipeline::{
-        AsyncEngine, AsyncEngineContext, AsyncEngineContextProvider, Error, ManyOut, PushRouter,
-        ResponseStream, SingleIn, async_trait,
+        AsyncEngine, AsyncEngineContextProvider, Error, ManyOut, PushRouter, ResponseStream,
+        SingleIn, async_trait,
     },
     protocols::annotated::Annotated,
 };
@@ -26,111 +19,25 @@ use tracing::Instrument;
 
 use crate::{
     kv_router::{
-        FindBestMatchOutcome, KvRouter,
-        metrics::RouterRequestMetrics,
-        sticky::coordinator::{StickySessionCoordinator, sticky_allowed_for_phase},
+        KvRouter, metrics::RouterRequestMetrics, sticky::coordinator::StickySessionCoordinator,
     },
     preprocessor::PreprocessedRequest,
-    protocols::{
-        TokenIdType,
-        common::{llm_backend::LLMEngineOutput, preprocessor::RoutingHints, timing::RequestPhase},
-    },
+    protocols::common::{llm_backend::LLMEngineOutput, timing::RequestPhase},
 };
 
+mod cancellation;
 mod request_guard;
+mod selection;
 
+use cancellation::{cancel_on_stop, cancelled_error};
 use request_guard::RequestGuard;
+use selection::{RoutingRequestParts, WorkerSelection};
 
 pub struct KvPushRouter {
     inner: PushRouter<PreprocessedRequest, Annotated<LLMEngineOutput>>,
     pub chooser: Arc<KvRouter>,
     /// Sticky session routing. Lazily activated when requests carry session_control.
     pub(super) sticky: Arc<StickySessionCoordinator>,
-}
-
-/// Result of worker selection containing instance ID, dp_rank, and overlap amount.
-struct WorkerSelection {
-    instance_id: u64,
-    dp_rank: u32,
-    overlap_amount: u32,
-    effective_overlap_blocks: f64,
-    cached_tokens: usize,
-    routing_hashes: Option<RoutingDecisionHashes>,
-    /// Whether the scheduler is tracking this request (add_request or
-    /// find_best_match_details with update_states=true was called).
-    scheduler_tracked: bool,
-}
-
-// NOTE: In KV router mode, worker selection is DP-rank precise. A pinned
-// worker without a concrete dp_rank is invalid unless the worker owns exactly
-// one rank and can be resolved unambiguously. Rank 0 is a real rank, not an
-// unset sentinel. Do not coerce unresolved ranks to 0.
-fn resolve_pinned_worker_rank(
-    worker_id: WorkerId,
-    requested_dp_rank: Option<u32>,
-    unique_dp_rank: Option<u32>,
-) -> Result<WorkerWithDpRank, Error> {
-    let Some(dp_rank) = requested_dp_rank.or(unique_dp_rank) else {
-        return Err(anyhow::anyhow!(
-            "Pinned worker {worker_id} requires an explicit dp_rank because it has multiple or unknown DP ranks"
-        ));
-    };
-
-    Ok(WorkerWithDpRank::new(worker_id, dp_rank))
-}
-
-#[derive(Clone, Copy)]
-struct RoutingRequestParts<'a> {
-    token_ids: &'a [TokenIdType],
-    block_mm_infos: Option<&'a [Option<BlockExtraInfo>]>,
-}
-
-impl<'a> RoutingRequestParts<'a> {
-    fn new(request: &'a PreprocessedRequest) -> Self {
-        let (token_ids, block_mm_infos) = request.block_mm_routing_info();
-        Self {
-            token_ids,
-            block_mm_infos,
-        }
-    }
-}
-
-fn cancelled_error(context_id: &str) -> Error {
-    DynamoError::builder()
-        .error_type(DynamoErrorType::Cancelled)
-        .message(format!("Request {context_id} was cancelled"))
-        .build()
-        .into()
-}
-
-async fn cancel_on_stop<T>(
-    context: &dyn AsyncEngineContext,
-    context_id: &str,
-    operation: impl Future<Output = T>,
-) -> Result<T, Error> {
-    tokio::pin!(operation);
-    tokio::select! {
-        biased;
-
-        // Keep completed ownership-bearing results so their normal cleanup can run.
-        result = &mut operation => Ok(result),
-        _ = context.stopped() => Err(cancelled_error(context_id)),
-    }
-}
-
-struct BestMatchArgs<'a> {
-    context_id: &'a str,
-    routing_parts: RoutingRequestParts<'a>,
-    router_config_override: Option<&'a RouterConfigOverride>,
-    update_states: bool,
-    return_routing_hashes: bool,
-    lora_name: Option<String>,
-    priority_jump: f64,
-    expected_output_tokens: Option<u32>,
-    pinned_worker: Option<WorkerWithDpRank>,
-    allowed_worker_ids: Option<HashSet<WorkerId>>,
-    routing_constraints: RoutingConstraints,
-    scheduler_tracked: bool,
 }
 
 impl KvPushRouter {
@@ -151,244 +58,6 @@ impl KvPushRouter {
             chooser,
             sticky,
         }
-    }
-
-    async fn select_best_match(&self, args: BestMatchArgs<'_>) -> Result<WorkerSelection, Error> {
-        let outcome = self
-            .chooser
-            .find_best_match_details(
-                Some(args.context_id),
-                args.routing_parts.token_ids,
-                args.routing_parts.block_mm_infos,
-                args.router_config_override,
-                args.update_states,
-                args.return_routing_hashes,
-                args.lora_name,
-                args.priority_jump,
-                args.expected_output_tokens,
-                args.pinned_worker,
-                args.allowed_worker_ids,
-                args.routing_constraints,
-            )
-            .await?;
-
-        match outcome {
-            FindBestMatchOutcome::Routed {
-                worker,
-                overlap_blocks,
-                effective_overlap_blocks,
-                cached_tokens,
-                routing_hashes,
-            } => Ok(WorkerSelection {
-                instance_id: worker.worker_id,
-                dp_rank: worker.dp_rank,
-                overlap_amount: overlap_blocks,
-                effective_overlap_blocks,
-                cached_tokens,
-                routing_hashes,
-                scheduler_tracked: args.scheduler_tracked,
-            }),
-            FindBestMatchOutcome::Backpressure {
-                reason,
-                queued_isl_tokens,
-                max_queued_isl_tokens,
-            } => Err(DynamoError::builder()
-                .error_type(DynamoErrorType::ResourceExhausted)
-                .message(format!(
-                    "router backpressure: {reason:?} (queued_isl_tokens={queued_isl_tokens}, max_queued_isl_tokens={max_queued_isl_tokens:?})"
-                ))
-                .build()
-                .into()),
-        }
-    }
-
-    /// Select a worker for the request, either using an exact phase-specific pin
-    /// or by finding the best KV overlap match.
-    async fn select_worker(
-        &self,
-        context_id: &str,
-        request: &PreprocessedRequest,
-        routing_parts: RoutingRequestParts<'_>,
-        phase: RequestPhase,
-        is_query_only: bool,
-        sticky_worker: Option<WorkerWithDpRank>,
-    ) -> Result<WorkerSelection, Error> {
-        let _nvtx_select = dynamo_nvtx_range!("route.select_worker");
-        let routing = request.routing.as_ref();
-        let lora_name = routing.and_then(|r| r.lora_name.clone());
-        let priority_jump = routing.and_then(|r| r.priority_jump).unwrap_or(0.0);
-        let expected_output_tokens = routing.and_then(|r| r.expected_output_tokens);
-        let allowed_worker_ids = routing.and_then(|r| r.allowed_worker_ids.clone());
-        let return_routing_hashes =
-            !is_query_only && self.chooser.indexer().records_routing_decisions();
-        let routing_constraints = routing
-            .and_then(|r| r.routing_constraints.clone())
-            .unwrap_or_default();
-        let sticky_pin = sticky_worker.map(|worker| (worker.worker_id, Some(worker.dp_rank)));
-        let Some((pinned_worker_id, requested_dp_rank)) =
-            pinned_worker_hint(phase, routing).or(sticky_pin)
-        else {
-            let _nvtx_kv = dynamo_nvtx_range!("route.kv_match");
-            let selection = self
-                .select_best_match(BestMatchArgs {
-                    context_id,
-                    routing_parts,
-                    router_config_override: request.router_config_override.as_ref(),
-                    update_states: !is_query_only,
-                    return_routing_hashes,
-                    lora_name,
-                    priority_jump,
-                    expected_output_tokens,
-                    pinned_worker: None,
-                    allowed_worker_ids,
-                    routing_constraints: routing_constraints.clone(),
-                    scheduler_tracked: !is_query_only,
-                })
-                .await?;
-
-            if !is_query_only {
-                let total_blocks = routing_parts
-                    .token_ids
-                    .len()
-                    .div_ceil(self.chooser.block_size() as usize);
-                // tests/utils/router_logs.py parses the structured fields on this event.
-                tracing::debug!(
-                    request_id = %context_id,
-                    worker_id = selection.instance_id,
-                    dp_rank = selection.dp_rank,
-                    overlap_blocks = selection.overlap_amount,
-                    total_blocks = total_blocks,
-                    "[ROUTING] Best: worker_{} dp_rank={} with {}/{} blocks overlap",
-                    selection.instance_id,
-                    selection.dp_rank,
-                    selection.overlap_amount,
-                    total_blocks,
-                );
-            }
-
-            return Ok(selection);
-        };
-
-        let pinned_worker = resolve_pinned_worker_rank(
-            pinned_worker_id,
-            requested_dp_rank,
-            self.chooser.unique_dp_rank_for_worker(pinned_worker_id),
-        )?;
-        {
-            let configs = self.chooser.workers_with_configs.borrow();
-            let eligibility = RoutingEligibility::new(
-                allowed_worker_ids.as_ref(),
-                None,
-                Some(pinned_worker),
-                &routing_constraints,
-            );
-            if let Err(error) = eligibility.validate_worker_rank(&configs, pinned_worker) {
-                return Err(anyhow::anyhow!(
-                    "Pinned worker {} dp_rank {} is not eligible: {error}",
-                    pinned_worker.worker_id,
-                    pinned_worker.dp_rank
-                ));
-            }
-        }
-
-        tracing::debug!(
-            worker_id = pinned_worker.worker_id,
-            dp_rank = pinned_worker.dp_rank,
-            ?phase,
-            "Routing to specified worker"
-        );
-
-        self.select_best_match(BestMatchArgs {
-            context_id,
-            routing_parts,
-            router_config_override: request.router_config_override.as_ref(),
-            update_states: !is_query_only,
-            return_routing_hashes,
-            lora_name,
-            priority_jump,
-            expected_output_tokens,
-            pinned_worker: Some(pinned_worker),
-            allowed_worker_ids,
-            routing_constraints,
-            scheduler_tracked: !is_query_only,
-        })
-        .await
-    }
-
-    fn sticky_worker_ineligibility_for_phase(
-        &self,
-        request: &PreprocessedRequest,
-        phase: RequestPhase,
-        worker: WorkerWithDpRank,
-    ) -> Option<WorkerEligibilityError> {
-        let routing = request.routing.as_ref()?;
-        if !sticky_allowed_for_phase(phase, Some(routing)) {
-            return None;
-        }
-
-        let default_constraints = RoutingConstraints::default();
-        let routing_constraints = routing
-            .routing_constraints
-            .as_ref()
-            .unwrap_or(&default_constraints);
-        let configs = self.chooser.workers_with_configs.borrow();
-        let eligibility = RoutingEligibility::new(
-            routing.allowed_worker_ids.as_ref(),
-            None,
-            Some(worker),
-            routing_constraints,
-        );
-        eligibility.validate_worker_rank(&configs, worker).err()
-    }
-
-    pub(crate) fn unbind_ineligible_sticky_worker_for_phase(
-        &self,
-        context_id: &str,
-        request: &PreprocessedRequest,
-        phase: RequestPhase,
-        worker: WorkerWithDpRank,
-    ) -> bool {
-        let Some(reason) = self.sticky_worker_ineligibility_for_phase(request, phase, worker)
-        else {
-            return false;
-        };
-
-        let Some((session_id, _binding)) = self.sticky.unbind_for_phase(request, phase) else {
-            return false;
-        };
-        tracing::warn!(
-            request_id = %context_id,
-            %session_id,
-            worker_id = worker.worker_id,
-            dp_rank = worker.dp_rank,
-            reason = %reason,
-            "Sticky worker is no longer eligible; removing session affinity"
-        );
-        true
-    }
-
-    pub(crate) async fn validate_sticky_worker_for_phase(
-        &self,
-        context_id: &str,
-        request: &PreprocessedRequest,
-        phase: RequestPhase,
-        worker: WorkerWithDpRank,
-    ) -> Result<WorkerWithDpRank, Error> {
-        let routing_parts = RoutingRequestParts::new(request);
-        let selection = self
-            .select_worker(
-                context_id,
-                request,
-                routing_parts,
-                phase,
-                true,
-                Some(worker),
-            )
-            .await?;
-        Ok(WorkerWithDpRank::new(
-            selection.instance_id,
-            selection.dp_rank,
-        ))
     }
 }
 
@@ -697,35 +366,6 @@ impl AsyncEngine<SingleIn<PreprocessedRequest>, ManyOut<Annotated<LLMEngineOutpu
     }
 }
 
-/// Extract a phase-specific (worker_id, dp_rank) pin from routing hints.
-///
-/// Returns `Some((worker_id, optional_dp_rank))` when the request should be
-/// pinned to a particular worker, or `None` when the normal KV-overlap
-/// selection path should be used.
-fn pinned_worker_hint(
-    phase: RequestPhase,
-    routing: Option<&RoutingHints>,
-) -> Option<(u64, Option<u32>)> {
-    let routing = routing?;
-    match phase {
-        RequestPhase::Prefill => {
-            let worker_id = routing.prefill_worker_id.or(routing.backend_instance_id)?;
-            let dp_rank = routing.prefill_dp_rank.or(routing.dp_rank);
-            Some((worker_id, dp_rank))
-        }
-        RequestPhase::Decode => {
-            let worker_id = routing.decode_worker_id.or(routing.backend_instance_id)?;
-            let dp_rank = routing.dp_rank;
-            Some((worker_id, dp_rank))
-        }
-        RequestPhase::Aggregated => {
-            let worker_id = routing.backend_instance_id?;
-            let dp_rank = routing.dp_rank;
-            Some((worker_id, dp_rank))
-        }
-    }
-}
-
 /// A direct routing wrapper for `RouterMode::Direct`.
 ///
 /// This wraps a `PushRouter` and reads worker IDs from each request's routing hints,
@@ -768,149 +408,5 @@ impl AsyncEngine<SingleIn<PreprocessedRequest>, ManyOut<Annotated<LLMEngineOutpu
         tracing::debug!(worker_id = worker_id, "Direct routing to specified worker");
 
         self.inner.direct(request, worker_id).await
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use std::{
-        collections::{HashMap, HashSet},
-        future::Future,
-        pin::Pin,
-        sync::{
-            Arc,
-            atomic::{AtomicBool, Ordering},
-        },
-        task::{Context, Poll},
-    };
-
-    use dynamo_kv_router::{
-        protocols::{RoutingConstraints, WorkerWithDpRank},
-        scheduling::{RoutingEligibility, WorkerEligibilityError},
-    };
-    use dynamo_runtime::{
-        error::{DynamoError, ErrorType},
-        pipeline::{AsyncEngineContext, context::Controller},
-    };
-
-    use super::{cancel_on_stop, pinned_worker_hint, resolve_pinned_worker_rank};
-    use crate::local_model::runtime_config::ModelRuntimeConfig;
-    use crate::protocols::common::{preprocessor::RoutingHints, timing::RequestPhase};
-
-    struct PendingUntilDropped(Arc<AtomicBool>);
-
-    impl Future for PendingUntilDropped {
-        type Output = ();
-
-        fn poll(self: Pin<&mut Self>, _cx: &mut Context<'_>) -> Poll<Self::Output> {
-            Poll::Pending
-        }
-    }
-
-    impl Drop for PendingUntilDropped {
-        fn drop(&mut self) {
-            self.0.store(true, Ordering::SeqCst);
-        }
-    }
-
-    #[tokio::test]
-    async fn cancel_on_stop_drops_pending_operation() {
-        let context = Controller::new("cancelled-request".to_string());
-        context.stop();
-        let dropped = Arc::new(AtomicBool::new(false));
-
-        let error = cancel_on_stop(&context, context.id(), PendingUntilDropped(dropped.clone()))
-            .await
-            .unwrap_err();
-
-        let error = error
-            .downcast_ref::<DynamoError>()
-            .expect("cancellation should return DynamoError");
-        assert_eq!(error.error_type(), ErrorType::Cancelled);
-        assert!(dropped.load(Ordering::SeqCst));
-    }
-
-    #[test]
-    fn resolve_pinned_worker_rank_uses_explicit_rank_including_zero() {
-        let worker = resolve_pinned_worker_rank(7, Some(0), Some(3)).unwrap();
-        assert_eq!(worker.worker_id, 7);
-        assert_eq!(worker.dp_rank, 0);
-    }
-
-    #[test]
-    fn resolve_pinned_worker_rank_uses_unique_rank_when_unset() {
-        let worker = resolve_pinned_worker_rank(7, None, Some(3)).unwrap();
-        assert_eq!(worker.worker_id, 7);
-        assert_eq!(worker.dp_rank, 3);
-    }
-
-    #[test]
-    fn resolve_pinned_worker_rank_rejects_unresolved_rank() {
-        let error = resolve_pinned_worker_rank(7, None, None)
-            .unwrap_err()
-            .to_string();
-        assert!(error.contains("requires an explicit dp_rank"));
-    }
-
-    #[test]
-    fn pinned_worker_hint_prefill_uses_prefill_worker_before_backend() {
-        let routing = RoutingHints {
-            backend_instance_id: Some(1),
-            prefill_worker_id: Some(2),
-            dp_rank: Some(3),
-            prefill_dp_rank: Some(4),
-            ..Default::default()
-        };
-
-        let hint = pinned_worker_hint(RequestPhase::Prefill, Some(&routing));
-        assert_eq!(hint, Some((2, Some(4))));
-    }
-
-    #[test]
-    fn pinned_worker_hint_decode_uses_decode_worker_before_backend() {
-        let routing = RoutingHints {
-            backend_instance_id: Some(1),
-            decode_worker_id: Some(5),
-            dp_rank: Some(6),
-            ..Default::default()
-        };
-
-        let hint = pinned_worker_hint(RequestPhase::Decode, Some(&routing));
-        assert_eq!(hint, Some((5, Some(6))));
-    }
-
-    #[test]
-    fn pinned_worker_hint_aggregated_uses_backend_worker() {
-        let routing = RoutingHints {
-            backend_instance_id: Some(9),
-            dp_rank: Some(7),
-            ..Default::default()
-        };
-
-        let hint = pinned_worker_hint(RequestPhase::Aggregated, Some(&routing));
-        assert_eq!(hint, Some((9, Some(7))));
-    }
-
-    #[test]
-    fn sticky_validation_style_ignores_transient_overload() {
-        let worker = WorkerWithDpRank::new(7, 0);
-        let configs = HashMap::from([(7, ModelRuntimeConfig::default())]);
-        let constraints = RoutingConstraints::default();
-        let overloaded = HashSet::from([7]);
-        let scheduling_eligibility =
-            RoutingEligibility::new(None, Some(&overloaded), Some(worker), &constraints);
-        let sticky_eligibility = RoutingEligibility::new(None, None, Some(worker), &constraints);
-
-        assert_eq!(
-            scheduling_eligibility
-                .validate_worker_rank(&configs, worker)
-                .err(),
-            Some(WorkerEligibilityError::WorkerOverloaded { worker_id: 7 })
-        );
-        assert!(
-            sticky_eligibility
-                .validate_worker_rank(&configs, worker)
-                .is_ok()
-        );
     }
 }
